+++ src/middleware/db.ts
@@ -1,17 +1,18 @@
 import { DATABASE_URL } from "@config/constants";
 import { drizzle } from "drizzle-orm/node-postgres";
 import { Context, Next } from "hono";
-import { Pool } from "pg";
 import * as auth from "../db/schema/auth";
 import * as videos from "../db/schema/videos";
 import * as tenants from "../db/schema/tenants";
+import pg from "pg";
+const { Pool } = pg;
+
+const pool = new Pool({
+  connectionString: DATABASE_URL,
+});
 
 export const dbMiddleware = async (c: Context, next: Next) => {
   const userId = c.get("userId");
-
-  const pool = new Pool({
-    connectionString: DATABASE_URL,
-  });
 
   const db = drizzle({
     client: pool,
@@ -22,6 +23,8 @@ export const dbMiddleware = async (c: Context, next: Next) => {
   if (userId) {
     await db.execute(`SET LOCAL app.user_id = ${userId}`);
   }
+
+  c.set("db", db);
 
   await next();
 };
